+++ code/src/sixsq/nuvla/server/resources/credential/vpn_utils.clj
@@ -2,6 +2,7 @@
   (:require
     [clj-http.client :as http]
     [clojure.data.json :as json]
+    [clojure.tools.logging :as log]
     [sixsq.nuvla.db.filter.parser :as parser]
     [sixsq.nuvla.server.resources.common.crud :as crud]
     [sixsq.nuvla.server.resources.configuration :as configuration]
@@ -9,8 +10,7 @@
     [sixsq.nuvla.server.resources.credential-template-infrastructure-service-vpn-customer
      :as tpl-customer]
     [sixsq.nuvla.server.util.log :as logu]
-    [sixsq.nuvla.server.util.response :as r]
-    [clojure.tools.logging :as log]))
+    [sixsq.nuvla.server.util.response :as r]))
 
 
 (defn get-service