+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Change
 
+  - NuvlaBox - update existing resources on commission
   - NuvlaBox - make NuvlaBox visible for vpn infrastructure when selected
   - Credential openvpn - change error message when duplicated credential
   - Update to parent 6.7.0